requirements.txt
anthropic==0.18.1
python-dotenv==1.0.1 
httpx==0.27.2
requests==2.31.0
beautifulsoup4==4.12.3
